browser clear v2

+++ src/services/submissionService.ts
@@ -12,7 +12,9 @@ function clearBrowserCache() {
   if (typeof indexedDB !== "undefined") {
     indexedDB.databases?.().then((databases) => {
       databases.forEach((db) => {
-        indexedDB.deleteDatabase(db.name);
+        if (db.name) {
+          indexedDB.deleteDatabase(db.name);
+        }
       });
     });
   }
